# front/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Le Renard Étudiant — API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/utilisateurs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Lister Utilisateurs",
        "operationId": "lister_utilisateurs_api_admin_utilisateurs_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 254,
              "description": "Début du pseudo ou de l'email",
              "default": "",
              "title": "Q"
            },
            "description": "Début du pseudo ou de l'email"
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "decalage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Decalage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageUtilisateurs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/utilisateurs/{utilisateur_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Modifier Utilisateur",
        "operationId": "modifier_utilisateur_api_admin_utilisateurs__utilisateur_id__patch",
        "parameters": [
          {
            "name": "utilisateur_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Utilisateur Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificationUtilisateur"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModificationUtilisateur": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "suspendu": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suspendu"
          }
        },
        "type": "object",
        "title": "ModificationUtilisateur"
      },
...
      "PageUtilisateurs": {
        "properties": {
          "utilisateurs": {
            "items": {
              "$ref": "#/components/schemas/UtilisateurAdmin"
            },
            "type": "array",
            "title": "Utilisateurs"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "utilisateurs",
          "total"
        ],
        "title": "PageUtilisateurs"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "apprenant",
          "contributeur",
          "relecteur",
          "admin"
        ],
        "title": "Role",
        "description": "Rôles du cadrage, du moins au plus élevé. Chaque rôle a les droits des précédents."
      },
...
      "UtilisateurAdmin": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "pseudo": {
            "type": "string",
            "title": "Pseudo"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "email_verifie": {
            "type": "boolean",
            "title": "Email Verifie"
          },
          "suspendu": {
            "type": "boolean",
            "title": "Suspendu"
          },
          "cree_le": {
            "type": "string",
            "format": "date-time",
            "title": "Cree Le"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "pseudo",
          "role",
          "email_verifie",
          "suspendu",
          "cree_le"
        ],
        "title": "UtilisateurAdmin"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}